refactor: hook form fields styles in separate file
