+++ components/HookForm/HFPassField/HFPassField.styles.ts
@@ -1,0 +1,8 @@
+export const sx = {
+  divider: {
+    height: 28,
+    marginRight: '14px',
+  },
+};
+
+export default sx;
