+++ components/HookForm/HFSendField/HFSendField.styles.ts
@@ -1,0 +1,13 @@
+export const sx = {
+  divider: {
+    height: 28,
+    marginRight: '14px',
+  },
+  adornment: {
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+};
+
+export default sx;
