+++ components/PA/Credentials/PasswordChangeModal/PasswordChangeModal.tsx
@@ -4,7 +4,7 @@ import { FormProvider, useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { Stack } from '@mui/material';
 
-import { HFPassField } from 'components/HookForm/HFPassField';
+import { HFPassField } from 'components/HookForm/HFPassField/HFPassField';
 import { Box } from 'components/UI/Box/Box';
 import { Modal } from 'components/UI/Modal/Modal';
 import { Typography } from 'components/UI/Typography/Typography';
