+++ pages/_document.tsx
@@ -11,11 +11,9 @@ class MyDocument extends Document {
             href='https://fonts.googleapis.com/css2?family=Roboto+Slab:wght@400;500;700&display=swap'
             rel='stylesheet'
           />
-
           <link rel='preconnect' href='https://fonts.googleapis.com' />
           <link rel='preconnect' href='https://fonts.gstatic.com' />
           <link href='https://fonts.googleapis.com/css2?family=Nunito:wght@400;500;700&display=swap' rel='stylesheet' />
-
           <link rel='preconnect' href='https://fonts.googleapis.com' />
           <link rel='preconnect' href='https://fonts.gstatic.com' crossOrigin='true' />
           <link href='https://fonts.googleapis.com/css2?family=Coming+Soon&display=swap' rel='stylesheet' />
